requirements/base.txt
celery==5.2.7
Django==3.2.19
dj-database-url==1.2.0
django-model-utils==4.3.1
django-cors-headers==3.13.0
django-avatar==7.1.1
django-jinja==2.10.2
django-utils-six==2.0
jinja2==3.1.2
django-filter==22.1
django-crispy-forms==1.14.0
easy-thumbnails==2.8.5
djoser==2.2.0
django-autoslug==1.9.8
django-localflavor==3.1
django-xhtml2pdf==0.0.4
django-fsm==2.8.1
django-logentry-admin==1.1.0
django-embed-video==1.4.8
django-ordered-model==3.6
django-activity-stream==1.4.2
django-picklefield==3.1

# sending emails
djmail==2.0.0

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-expander==0.2.3
drf-extra-fields==3.4.1
djangorestframework-timed-auth-token==1.3.0

# runs the app server
gunicorn==20.1.0

# postgresql driver
psycopg2==2.9.5

# payments
stripe==1.35.0

# Django celery results
django_celery_results==2.4.0

# other stuffs
# git+https://github.com/jmichelsen/AWeber-API-Python-Library.git@update_library_to_python3.6 # AWeber
aweber-api @ git+https://github.com/jmichelsen/AWeber-API-Python-Library.git@c88876e4624702a251b881be3f8e6e62d48aad9a
WeasyPrint==57.2
pytz==2022.7.1
xhtml2pdf==0.2.9
img-rotate==0.1.1  # fix image orientations
python-dateutil==2.8.2
html5lib==1.1
pdfrw==0.4
pypdf==3.1.0
pikepdf==6.2.9

# BeautifulSoup
beautifulsoup4==4.11.2

# SMS
twilio==7.16.2
kombu==5.2.4

# SFTP connection
pysftp==0.2.9

# Phone numbers
phonenumbers==8.13.5

# configs
django-constance[database]==2.9.1

# SendGrid bindings for Python
sendgrid==6.9.7

# Virtru SDK
virtru-sdk==2.4.0

# Maps
googlemaps==4.10.0

reportlab==3.6.13
